Refactor plots and data to separate modules
